refactor: fixed unintentionally overwrite of user to existing user and re-structure authentication and authorization on NotificationService

+++ elms-api/app/Services/notifications/NotificationService.php
@@ -24,30 +24,6 @@ class NotificationService {
     public function store(int $user_id, string $title, string $type, string $message, bool $is_read = false, ?array $data = null): void {
 
         if($this->current_user_id) {
-            
-            $assigned_to = $this->db->query("SELECT * FROM users WHERE id = :id", [
-                'id' => $this->current_user_id,
-            ])->find();
-
-            if(!$assigned_to) {
-                http_response_code(404);
-                echo json_encode([
-                    "success" => false,
-                    "message" => "Assigned employee not found",
-                    "user_id" => $user_id,
-                ]);
-                return;
-            }
-            
-            if($assigned_to['id'] !== $user_id) {
-                http_response_code(403);
-                echo json_encode([
-                    "success" => false,
-                    "message" => "You are not authorized to create a notification for this employee",
-                    "user_id" => $user_id,
-                ]);
-                return;
-            }
 
             $this->db->query("INSERT INTO notifications (user_id, title, message, type, read_at, data) VALUES (:user_id, :title, :message, :type, :read_at, :data)", [
                 'user_id' => $user_id, // recipient id (who will receive the notification)
@@ -57,6 +33,14 @@ class NotificationService {
                 'read_at' => $is_read ? date('Y-m-d H:i:s') : null,
                 'data' => $data ? json_encode($data) : null,
             ]);
+
+            http_response_code(201);
+            echo json_encode([
+                'success' => true,
+                'message' => 'Notification created successfully',
+                'user_id' => $user_id,
+            ]);
+            return;
         }
 
         http_response_code(401);
@@ -74,21 +58,30 @@ class NotificationService {
      */
     public function index() {
 
-        if(!$this->current_user_id) {
-            http_response_code(401);
+        if($this->current_user_id) {
+
+            $notifications = $this->db->query("SELECT * FROM notifications WHERE user_id = :user_id ORDER BY created_at DESC", [
+            'user_id' => $this->current_user_id, // list of notifications for the current user
+            ])->all();
+
+            http_response_code(200);
             echo json_encode([
-                'success' => false,
-                'message' => 'Unathorized Access',
-                'user_id' => $this->current_user_id,
+                'success' => true,
+                'message' => 'Notifications fetched successfully',
+                'notifications' => $notifications,
             ]);
-            return;
+            return $notifications ?? [];
         }
 
-        $notifications = $this->db->query("SELECT * FROM notifications WHERE user_id = :user_id ORDER BY created_at DESC", [
-            'user_id' => $this->current_user_id, // list of notifications for the current user
-        ])->all();
+        http_response_code(401);
+        echo json_encode([
+            'success' => false,
+            'message' => 'Unathorized Access',
+            'user_id' => $this->current_user_id,
+        ]);
+        return;
 
-        return $notifications ?? [];
+       
     }
 
 
@@ -108,10 +101,35 @@ class NotificationService {
             return false;
         }
 
-        $this->db->query("UPDATE notifications SET read_at = :read_at WHERE id = :id", [
+        $notification = $this->db->query("SELECT * FROM notifications WHERE id = :id AND user_id = :user_id", [
+            'id' => $id,
+            'user_id' => $this->current_user_id,
+        ])->find();
+
+        if(!$notification) {
+            http_response_code(404);
+            echo json_encode([
+                'success' => false,
+                'message' => 'Notification not found',
+                'user_id' => $this->current_user_id,
+            ]);
+            return false;
+        }
+        if($notification['read_at'] == date('Y-m-d H:i:s')) {
+            http_response_code(400);
+            echo json_encode([
+                'success' => false,
+                'message' => 'Notification already marked as read',
+                'user_id' => $this->current_user_id,
+            ]);
+            return false;
+        }
+
+        $this->db->query("UPDATE notifications SET read_at = :read_at WHERE id = :id AND user_id = :user_id", [
             'read_at' => date('Y-m-d H:i:s'),
             'id' => $id,
-        ]);
+            'user_id' => $this->current_user_id,
+        ])->find();
 
         http_response_code(200);
         echo json_encode([
@@ -139,8 +157,9 @@ class NotificationService {
             return;
         }
 
-        $this->db->query("DELETE FROM notifications WHERE id = :id", [
+        $this->db->query("DELETE FROM notifications WHERE id = :id AND user_id = :user_id", [
             'id' => $id,
+            'user_id' => $this->current_user_id,
         ])->find();
 
         http_response_code(200);
